Moved selectMoodFrequency to actions/data-visualization
